+++ beacon-chain/cache/depositcache/BUILD.bazel
@@ -1,0 +1,29 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
+
+go_library(
+    name = "go_default_library",
+    srcs = [
+        "deposits_cache.go",
+        "pending_deposits.go",
+    ],
+    importpath = "github.com/prysmaticlabs/prysm/beacon-chain/cache/depositcache",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//proto/eth/v1alpha1:go_default_library",
+        "//shared/hashutil:go_default_library",
+        "@com_github_prometheus_client_golang//prometheus:go_default_library",
+        "@com_github_prometheus_client_golang//prometheus/promauto:go_default_library",
+        "@com_github_sirupsen_logrus//:go_default_library",
+        "@io_opencensus_go//trace:go_default_library",
+    ],
+)
+
+go_test(
+    name = "go_default_test",
+    srcs = ["pending_deposits_test.go"],
+    embed = [":go_default_library"],
+    deps = [
+        "//proto/eth/v1alpha1:go_default_library",
+        "@com_github_gogo_protobuf//proto:go_default_library",
+    ],
+)
