+++ beacon-chain/cache/depositcache/pending_deposits_test.go
@@ -1,4 +1,4 @@
-package db
+package depositcache
 
 import (
 	"context"
