+++ beacon-chain/db/BUILD.bazel
@@ -8,8 +8,6 @@ go_library(
         "block_operations.go",
         "db.go",
         "deposit_contract.go",
-        "deposits_cache.go",
-        "pending_deposits.go",
         "schema.go",
         "setup_db.go",
         "state.go",
@@ -19,6 +17,7 @@ go_library(
     importpath = "github.com/prysmaticlabs/prysm/beacon-chain/db",
     visibility = ["//beacon-chain:__subpackages__"],
     deps = [
+        "//beacon-chain/cache/depositcache:go_default_library",
         "//beacon-chain/core/blocks:go_default_library",
         "//beacon-chain/core/state:go_default_library",
         "//proto/beacon/p2p/v1:go_default_library",
@@ -47,7 +46,6 @@ go_test(
         "block_test.go",
         "db_test.go",
         "deposit_contract_test.go",
-        "pending_deposits_test.go",
         "state_test.go",
         "validator_test.go",
     ],
